+++ Prowl.Runtime/GUI/Widgets/Gui.InputField.cs
@@ -102,7 +102,7 @@ namespace Prowl.Runtime.GUI
                         render_pos.y -= g.CurrentNode.VScroll;
 
                     Color32 colb = style.TextColor;
-                    g.Draw2D.DrawList.AddText(font, (float)fontsize, render_pos, colb, value, 0, value.Length, 0.0f, null);
+                    g.Draw2D.DrawList.AddText(font, fontsize, render_pos, colb, value, 0, value.Length, 0.0f, null);
                 }
                 g.Draw2D.PopClip();
 
@@ -265,7 +265,7 @@ namespace Prowl.Runtime.GUI
             if ((Flags & InputFieldFlags.OnlyDisplay) == InputFieldFlags.OnlyDisplay)
             {
                 Color32 colb = style.TextColor;
-                g.Draw2D.DrawList.AddText(font, (float)fontsize, render_pos - render_scroll, colb, stb.Text, 0, stb.Text.Length, 0.0f, (is_multiline ? null : (Vector4?)clip_rect));
+                g.Draw2D.DrawList.AddText(font, fontsize, render_pos - render_scroll, colb, stb.Text, 0, stb.Text.Length, 0.0f, (is_multiline ? null : (Vector4?)clip_rect));
                 return false;
             }
 
@@ -307,7 +307,7 @@ namespace Prowl.Runtime.GUI
 
 
             Color32 col = style.TextColor;
-            g.Draw2D.DrawList.AddText(font, (float)fontsize, render_pos - render_scroll, col, stb.Text, 0, stb.Text.Length, 0.0f, (is_multiline ? null : (Vector4?)clip_rect));
+            g.Draw2D.DrawList.AddText(font, fontsize, render_pos - render_scroll, col, stb.Text, 0, stb.Text.Length, 0.0f, (is_multiline ? null : (Vector4?)clip_rect));
             //g.DrawText(font, fontsize, Text, render_pos - render_scroll, Color.black, 0, stb.CurLenA, 0.0f, (is_multiline ? null : (ImVec4?)clip_rect));
 
             // Draw blinking cursor
